CR Feedback: moving utils out of common and into feature;
